requirements.txt
aniso8601==8.1.0
bcrypt==3.1.7
blessings==1.7
bpython==0.20.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
curtsies==0.3.4
docker==4.4.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.1
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
jellyfish==0.8.2
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.24.1
mypy-extensions==0.4.3
Pillow==7.0.0
prettytable==2.0.0
prompt-toolkit==3.0.9
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.7.3
python-dotenv==0.15.0
python-nubia==0.2b5
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
simplejson==3.16.0
six==1.15.0
SQLAlchemy==1.3.12
termcolor==1.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
